refactor(git): split repository.rs god module into submodules
repository.rs was 1841 lines interleaving five responsibilities. Split
into a repository/ directory: worktree, branch, status, ci, paths — with
mod.rs as the root that re-exports the entire public API unchanged
(okena_git::repository::* and okena_git::* paths are byte-for-byte the
same) and holds shared private helpers + a test_support module.

Pure move: 104 tests pass identically, no logic or API change. Updates
the stale okena-git/CLAUDE.md Files table to the new layout.
